+++ SpeedyWeather/src/dynamics/forcing.jl
@@ -242,12 +242,26 @@ function forcing!(
         lf::Integer,
         model::AbstractModel,
     )
+    (; latds) = model.geometry
     # scale by radius as is the vorticity equation
     s = forcing.strength * vars.prognostic.scale[]
     k = forcing.wavenumber
 
     Fu = vars.tendencies.grid.u
-    return set!(Fu, (λ, θ, σ) -> s * sind(k * θ), model.geometry)
+    launch!(
+        architecture(Fu), RingGridWorkOrder, size(Fu), kolmogorov_flow_kernel!,
+        Fu, s, k, latds
+    )
+    return nothing
+end
+
+@kernel inbounds = true function kolmogorov_flow_kernel!(Fu, s, k, latds)
+    I = @index(Global, NTuple)
+    ij = I[1]
+    layer = I[2]
+    lat = latds[ij]
+
+    Fu[ij, layer] += s * sin(deg2rad(k * lat))
 end
 
 export HeldSuarez