+++ SpeedyWeather/src/models/primitive_dry.jl
@@ -231,8 +231,7 @@ function initialize!(model::PrimitiveDry; time::DateTime = DEFAULT_DATE)
 
     # set the time first
     (; clock) = variables.prognostic
-    clock.time = time       # set the current time
-    clock.start = time      # and store the start time
+    set!(clock, time = time, start = time)
 
     # set all initial conditions for the ocean, seaice, land then atmosphere
     initialize!(variables, model)