+++ SpeedyWeather/src/output/feedback.jl
@@ -83,6 +83,9 @@ function initialize!(feedback::Feedback, clock::Clock, model::AbstractModel)
     return nothing
 end
 
+# fallback if feedback is set to nothing
+initialize!(::Nothing, clock::Clock, model::AbstractModel) = nothing
+
 progress!(feedback::Feedback) = ProgressMeter.next!(feedback.progress_meter)
 
 function progress!(feedback::Feedback, vars::Variables)
@@ -96,10 +99,15 @@ function progress!(feedback::Feedback, vars::Variables)
     return nothing
 end
 
+progress!(::Nothing, progn::PrognosticVariables, diagn::DiagnosticVariables) = nothing
+
 """
 $(TYPEDSIGNATURES)
 Finalises the progress meter and the progress txt file."""
 finalize!(F::Feedback) = ProgressMeter.finish!(F.progress_meter)
+
+# fallback if feedback is set to nothing
+finalize!(::Nothing) = nothing
 
 """$(TYPEDSIGNATURES)
 Detect NaN (Not-a-Number, or Inf) in the prognostic variables."""
@@ -144,9 +152,9 @@ end
 # constant from the ProgressMeter module
 const FEEDBACK_DT_IN_SEC = Ref(1.0)
 const FEEDBACK_TIME = Ref(DEFAULT_DATE)
-const FEEDBACK_UMAX = Ref(-1f0)     # default negative = skip show
-const FEEDBACK_TMIN = Ref(-1f0)
-const FEEDBACK_TMAX = Ref(-1f0)
+const FEEDBACK_UMAX = Ref(-1.0f0)     # default negative = skip show
+const FEEDBACK_TMIN = Ref(-1.0f0)
+const FEEDBACK_TMAX = Ref(-1.0f0)
 
 # "extend" the speedstring function from ProgressMeter by defining it for ::AbstractFloat
 # not just ::Any to effectively overwrite it
@@ -279,6 +287,7 @@ Writes the time stepping progress to the progress.txt file every `every_n_percen
 function callback!(progress_txt::ProgressTxt, vars, model)
     # escape in case of no output
     progress_txt.write_only_with_output && (model.output.active || return nothing)
+    isnothing(model.feedback) && return nothing
 
     (; progress_meter, nans_detected) = model.feedback
     (; counter, n) = progress_meter
@@ -308,6 +317,7 @@ Finalizes the ProgressTxt callback by writing the total time taken to the progre
 function finalize!(progress_txt::ProgressTxt, vars, model)
     # escape in case of no output
     progress_txt.write_only_with_output && (model.output.active || return nothing)
+    isnothing(model.feedback) && return nothing
 
     (; file) = progress_txt
     (; progress_meter) = model.feedback