+++ SpeedyWeather/src/output/jld2_output.jl
@@ -72,7 +72,6 @@ Initialize JLD2 `output` by creating a JLD2 file.
 To be called just before the first timesteps."""
 function initialize!(
         output::JLD2Output,
-        feedback::AbstractFeedback,
         vars::Variables,
         model::AbstractModel,
     )