+++ SpeedyWeather/src/output/netcdf_output.jl
@@ -21,6 +21,8 @@ $(TYPEDFIELDS)"""
         Field2D,
         Field3D,
         Interpolator,
+        DT,
+        S,
     } <: AbstractOutput
 
     # FILE OPTIONS
@@ -64,10 +66,10 @@ $(TYPEDFIELDS)"""
 
     # WHAT/WHEN OPTIONS
     "[DERIVD] start date of the simulation, used for time dimension in netcdf file"
-    startdate::DateTime = DateTime(2000, 1, 1)
+    startdate::DT = DateTime(2000, 1, 1)
 
     "[OPTION] output frequency, time step"
-    output_dt::Second = Second(DEFAULT_OUTPUT_DT)
+    output_dt::S = Second(DEFAULT_OUTPUT_DT)
 
     "[OPTION] dictionary of variables to output, e.g. u, v, vor, div, pres, temp, humid"
     variables::OUTPUT_VARIABLES_DICT = OutputVariablesDict()
@@ -241,12 +243,14 @@ function set!(output::AbstractOutput; active, reset_path = true)
     return nothing
 end
 
+# fallback for nothing output
+set!(::Nothing; active, reset_path = true) = nothing
+
 """$(TYPEDSIGNATURES)
 Initialize NetCDF `output` by creating a netCDF file and storing the initial conditions
 of `vars`. To be called just before the first timesteps."""
 function initialize!(
         output::NetCDFOutput,
-        feedback::AbstractFeedback,
         vars::Variables,
         model::AbstractModel,
     )
@@ -292,9 +296,10 @@ function initialize!(
     output!(output, vars.prognostic.clock.time)   # write initial time
 
     # DEFINE NETCDF DIMENSIONS SPACE
+    # explictly move to CPU to avoid issues with Reactant and on GPU
     lond = get_lond(output.field2D)
     latd = get_latd(output.field2D)
-    σ = model.geometry.σ_levels_full
+    σ = on_architecture(CPU(), model.geometry.σ_levels_full)
     soil_indices = collect(1:get_soil_layers(model))
 
     defVar(dataset, "lon", lond, ("lon",), attrib = Dict("units" => "degrees_east", "long_name" => "longitude"))
@@ -319,6 +324,9 @@ function initialize!(
     output.write_restart && add!(model.callbacks, :restart_file => RestartFile())
     return nothing
 end
+
+# fallback for nothing output
+initialize!(::Nothing, ::Union{AbstractFeedback, Nothing}, ::Variables, ::AbstractModel) = nothing
 
 Base.close(output::NetCDFOutput) = NCDatasets.close(output.netcdf_file)
 Base.close(::Nothing) = nothing     # in case of no netCDF output nothing to close
@@ -358,6 +366,11 @@ function output!(output::AbstractOutput, simulation::AbstractSimulation)
     output!(output, clock.time)                                         # increase counter write time
     output!(output, output.variables, simulation)                       # write variables
     return output
+end
+
+# fallback for nothing output
+function output!(::Nothing, ::AbstractSimulation)
+    return nothing
 end
 
 get_indices(i, variable::AbstractOutputVariable) = get_indices(i, Val.(variable.dims_xyzt)...)
@@ -460,6 +473,8 @@ function finalize!(
     return close(output)
 end
 
+finalize!(::Nothing, ::AbstractSimulation) = nothing
+
 # default finalize method for output variables
 function finalize!(
         output::AbstractOutput,
@@ -538,3 +553,13 @@ function load_trajectory(var_name::Union{Symbol, String}, model::AbstractModel)
     @assert model.output.active "Output is turned off"
     return Array(NCDataset(get_full_output_file_path(model.output))[string(var_name)])
 end
+
+"""
+$(TYPEDSIGNATURES)
+Returns the output time step of the model `M`."""
+function get_output_dt(output::AbstractOutput)
+    return output.output_dt
+end
+
+# Fallback for when output is nothing
+get_output_dt(::Nothing) = Millisecond(0)