+++ SpeedyWeather/src/parameterizations/ocean.jl
@@ -129,7 +129,7 @@ function timestep!(
     (; monthly_temperature) = ocean
     (; sea_surface_temperature) = vars.prognostic.ocean
     NF = eltype(sea_surface_temperature)
-    weight = convert(NF, Dates.days(time - Dates.firstdayofmonth(time)) / Dates.daysinmonth(time))
+    weight = convert(NF, Dates.days(time - Dates.firstdayofmonth(time)) / Dates.daysinmonth(year(time), Dates.month(time)))
 
     launch!(
         architecture(sea_surface_temperature), LinearWorkOrder, size(sea_surface_temperature),