+++ SpeedyWeather/test/GPU/runtests.jl
@@ -54,6 +54,8 @@ include("primitive_wet.jl")
 if gpu_backend === :CUDA
 
     include("CUDA/architecture.jl")
+    # REACTANT ON GPU (currently only works tested with CUDA)
+    #include("reactant.jl")
 
 elseif gpu_backend === :AMDGPU
 
@@ -62,3 +64,4 @@ elseif gpu_backend === :AMDGPU
 elseif gpu_backend === :Metal
     include("MetalGPU/metal.jl")
 end
+