+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Unreleased
 
 - [BREAKING] New dynamic variable system based on NamedTuples [#969](https://github.com/SpeedyWeather/SpeedyWeather.jl/pull/969)
+- Broadcasting for LTA and RingGrids revised [#977](https://github.com/SpeedyWeather/SpeedyWeather.jl/pull/977)
+- Reactant support for the BarotropicModel and Reactant architecture basics [#970](https://github.com/SpeedyWeather/SpeedyWeather.jl/pull/970)
 - Minor fixes for Metal GPUs [#1019](https://github.com/SpeedyWeather/SpeedyWeather.jl/pull/1019)
 - Delete unused/moved zenith.jl file [#1016](https://github.com/SpeedyWeather/SpeedyWeather.jl/pull/1016)
 - Add `get_output_path(::Simulation)` to retrieve the output file path [#1011](https://github.com/SpeedyWeather/SpeedyWeather.jl/pull/1011)